expose task extent as a selector
