+++ src/stores/tasks_selectors.js
@@ -1,4 +1,5 @@
 import { createSelector } from 'reselect';
+import d3 from 'd3';
 
 const tasksSelector = state => state.tasks;
 
@@ -30,6 +31,17 @@ const getTaskSummary = state => (
   getCurrentTask(state).value.stats
 );
 
+const getTaskExtent = state => {
+  const taskSummary = getTaskSummary(state);
+  const createdAt = taskSummary.date * 1000;
+
+  const dateFormat = d3.time.format('%Y-%m-%d');
+  const fromDate = dateFormat(new Date(createdAt));
+  const toDate = dateFormat(new Date());
+
+  return { fromDate, toDate };
+};
+
 const TasksSelectors = {
   getAllTasks: createSelector(tasksSelector, getAllTasks),
   getCompletedTasks: createSelector(tasksSelector, getCompletedTasks),
@@ -37,6 +49,7 @@ const TasksSelectors = {
   getCurrentTask: createSelector(tasksSelector, getCurrentTask),
   getCurrentTaskId: createSelector(tasksSelector, getCurrentTaskId),
   getTaskSummary: createSelector(tasksSelector, getTaskSummary),
+  getTaskExtent: createSelector(tasksSelector, getTaskExtent),
 };
 
 export default TasksSelectors;
